# Lammah-game-backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/games": {
      "post": {
        "operationId": "gamesCreate",
        "summary": "Create a new game with 2 teams",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGameDto"
              },
              "examples": {
                "default": {
                  "summary": "Create game",
                  "value": {
                    "name": "Friday Family Game",
                    "teams": [
                      {
                        "name": "Team Falcons",
                        "members": [
                          "Muath",
                          "Sara"
                        ]
                      },
                      {
                        "name": "Team Stars",
                        "members": [
                          "Noura",
                          "Fahad"
                        ]
                      }
                    ],
                    "categoryIds": [
                      "66b8f5f2c9d7a8b1e3f00222",
                      "66b8f5f2c9d7a8b1e3f00223"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Game created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameMutationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Exactly 2 teams are required",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Subscription required after free game is used",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "You need an active subscription to create more games.",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "gamesList",
        "summary": "Get all games",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Games retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/games/{id}": {
      "get": {
        "operationId": "gamesGetById",
        "summary": "Get a specific game by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Game MongoDB ObjectId",
            "schema": {
              "example": "66b8f5f2c9d7a8b1e3f00444",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Game retrieved successfully. Answers hidden by default.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetailResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Game not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 404,
                    "message": "Game with ID \"66b8f5f2c9d7a8b1e3f00444\" not found",
                    "error": "Not Found"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/games/{id}/reveal-answer": {
      "post": {
        "operationId": "gamesRevealAnswer",
        "summary": "Reveal the correct answer for a question",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Game MongoDB ObjectId",
            "schema": {
              "example": "66b8f5f2c9d7a8b1e3f00444",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevealAnswerDto"
              },
              "examples": {
                "default": {
                  "summary": "Reveal answer",
                  "value": {
                    "questionId": "66b8f5f2c9d7a8b1e3f00333"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Answer revealed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameMutationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - question not found in game",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Question not found in this game board",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Game not found"
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/games/{id}/award-points": {
      "post": {
        "operationId": "gamesAwardPoints",
        "summary": "Award points to a team for answering correctly",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Game MongoDB ObjectId",
            "schema": {
              "example": "66b8f5f2c9d7a8b1e3f00444",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "questionId",
                  "teamIndex"
                ],
                "properties": {
                  "questionId": {
                    "type": "string",
                    "example": "66b8f5f2c9d7a8b1e3f00333"
                  },
                  "teamIndex": {
                    "type": "number",
                    "example": 0
                  }
                }
              },
              "examples": {
                "default": {
                  "summary": "Award points to team 0",
                  "value": {
                    "questionId": "66b8f5f2c9d7a8b1e3f00333",
                    "teamIndex": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Points awarded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameMutationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid teamIndex or question already answered",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Question is already answered",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Game not found"
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/games/{id}/skip-question": {
      "post": {
        "operationId": "gamesSkipQuestion",
        "summary": "Skip a question without awarding points",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Game MongoDB ObjectId",
            "schema": {
              "example": "66b8f5f2c9d7a8b1e3f00444",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkipQuestionDto"
              },
              "examples": {
                "default": {
                  "summary": "Skip question",
                  "value": {
                    "questionId": "66b8f5f2c9d7a8b1e3f00333"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question skipped successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameMutationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - question already answered",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Question is already answered",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Game not found"
          }
        },
        "tags": [
          "Games"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Lammah Game API",
    "description": "Runtime HTTP contract for Lammah game administration, content, and gameplay.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CatalogLocalizedTextResponseDto": {
        "type": "object",
        "properties": {
          "ar": {
            "type": "string"
          },
          "en": {
            "type": "string"
          }
        },
        "required": [
          "ar",
          "en"
        ]
      },
...
      "CategoryCatalogResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/CatalogLocalizedTextResponseDto"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "name",
          "slug"
        ]
      },
      "CategoryBannerResponseDto": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "mimetype": {
            "type": "string"
          },
          "size": {
            "type": "number"
          }
        },
        "required": [
          "filename",
          "url",
          "mimetype",
          "size"
        ]
      },
      "CategoryAiConfigDto": {
        "type": "object",
        "properties": {
          "knowledgeFile": {
            "type": "string",
            "example": "sports/world-cup.md"
          },
          "temperature": {
            "type": "number",
            "example": 0.7,
            "minimum": 0,
            "maximum": 2
          },
          "preferredQuestionTypes": {
            "example": [
              "historical moments",
              "records",
              "finals"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "avoidTopics": {
            "example": [
              "winner of 2022",
              "host country 2022"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "extraInstructions": {
            "type": "string",
            "example": "Focus on fun offline party-game questions."
          }
        }
      },
      "CategoryGameplayConfigDto": {
        "type": "object",
        "properties": {
          "gameModes": {
            "type": "object",
            "example": {
              "trivia": 30,
              "identifyCharacter": 30,
              "identifyVoice": 20,
              "completeQuote": 20
            }
          },
          "questionTypes": {
            "type": "object",
            "description": "Legacy field. Prefer gameModes for new categories.",
            "example": {
              "text": 20,
              "image": 35,
              "audio": 25,
              "quote": 10,
              "emoji": 5,
              "timeline": 5
            }
          },
          "supportedAssetTypes": {
            "example": [
              "text",
              "image",
              "audio",
              "quote",
              "emoji",
              "timeline"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "musicConfig": {
            "type": "object",
            "example": {
              "allowedRegions": [
                "khaliji",
                "egyptian"
              ],
              "allowedLanguages": [
                "ar"
              ],
              "maxPreviewDuration": 15
            }
          },
          "preferredDifficultyMix": {
            "type": "object",
            "example": {
              "easy": 30,
              "medium": 50,
              "hard": 20
            }
          },
          "maxAudioDuration": {
            "type": "number",
            "example": 6,
            "minimum": 1,
            "maximum": 20
          },
          "imageRevealAllowed": {
            "type": "boolean",
            "example": true
          },
          "allowMultipleAssets": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "CategoryResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "catalogId": {
            "type": "string",
            "nullable": true
          },
          "catalog": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryCatalogResponseDto"
              }
            ]
          },
          "banner": {
            "$ref": "#/components/schemas/CategoryBannerResponseDto"
          },
          "aiConfig": {
            "$ref": "#/components/schemas/CategoryAiConfigDto"
          },
          "gameplayConfig": {
            "$ref": "#/components/schemas/CategoryGameplayConfigDto"
          },
          "isActive": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "name",
          "slug",
          "isActive",
          "sortOrder"
        ]
      },
...
      "TeamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "members": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "members"
        ]
      },
      "CreateGameDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "teams": {
            "minItems": 2,
            "maxItems": 2,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamDto"
            }
          },
          "categoryIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "teams",
          "categoryIds"
        ]
      },
      "GameTeamResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "members": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "score": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "members",
          "score"
        ]
      },
      "GameBoardQuestionResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "question": {
            "$ref": "#/components/schemas/QuestionResponseDto"
          },
          "points": {
            "type": "number",
            "enum": [
              200,
              400,
              600
            ]
          },
          "isAnswered": {
            "type": "boolean"
          },
          "isAnswerRevealed": {
            "type": "boolean"
          },
          "answeredByTeamIndex": {
            "type": "number"
          },
          "awardedPoints": {
            "type": "number"
          }
        },
        "required": [
          "question",
          "points",
          "isAnswered",
          "isAnswerRevealed"
        ]
      },
      "GameCategoryBoardResponseDto": {
        "type": "object",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/CategoryResponseDto"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameBoardQuestionResponseDto"
            }
          }
        },
        "required": [
          "category",
          "questions"
        ]
      },
      "GameResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "waiting",
              "active",
              "finished"
            ]
          },
          "teams": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameTeamResponseDto"
            }
          },
          "selectedCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryResponseDto"
            }
          },
          "board": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameCategoryBoardResponseDto"
            }
          },
          "currentTurnTeamIndex": {
            "type": "number"
          },
          "finishedAt": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "name",
          "status",
          "teams",
          "selectedCategories",
          "board",
          "currentTurnTeamIndex"
        ]
      },
      "GameMutationResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/GameResponseDto"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "data",
          "message"
        ]
      },
      "GameListResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameResponseDto"
            }
          }
        },
        "required": [
          "statusCode",
          "data"
        ]
      },
      "GameDetailResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/GameResponseDto"
          }
        },
        "required": [
          "statusCode",
          "data"
        ]
      },
      "RevealAnswerDto": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "string"
          }
        },
        "required": [
          "questionId"
        ]
      },
      "SkipQuestionDto": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "string"
          }
        },
        "required": [
          "questionId"
        ]
      },
...
    }
  }
}