# Lammah-game-frontend/src/api/generated/models/categoryResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lammah Game API
 * Runtime HTTP contract for Lammah game administration, content, and gameplay.
 * OpenAPI spec version: 1.0.0
 */
import type { CategoryResponseDtoCatalog } from "./categoryResponseDtoCatalog";
import type { CategoryBannerResponseDto } from "./categoryBannerResponseDto";
import type { CategoryAiConfigDto } from "./categoryAiConfigDto";
import type { CategoryGameplayConfigDto } from "./categoryGameplayConfigDto";

export interface CategoryResponseDto {
  _id: string;
  id?: string;
  name: string;
  slug: string;
  description?: string;
  /** @nullable */
  catalogId?: string | null;
  /** @nullable */
  catalog?: CategoryResponseDtoCatalog;
  banner?: CategoryBannerResponseDto;
  aiConfig?: CategoryAiConfigDto;
  gameplayConfig?: CategoryGameplayConfigDto;
  isActive: boolean;
  sortOrder: number;
  createdAt?: string;
  updatedAt?: string;
}