# Lammah-game-frontend/src/api/generated/models/gameBoardQuestionResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Lammah Game API
 * Runtime HTTP contract for Lammah game administration, content, and gameplay.
 * OpenAPI spec version: 1.0.0
 */
import type { QuestionResponseDto } from "./questionResponseDto";
import type { GameBoardQuestionResponseDtoPoints } from "./gameBoardQuestionResponseDtoPoints";

export interface GameBoardQuestionResponseDto {
  _id?: string;
  question: QuestionResponseDto;
  points: GameBoardQuestionResponseDtoPoints;
  isAnswered: boolean;
  isAnswerRevealed: boolean;
  answeredByTeamIndex?: number;
  awardedPoints?: number;
}